:truck: rename: Dto와 Doamin 파일 분리
