+++ build.gradle
@@ -34,6 +34,7 @@ dependencies {
 	implementation 'com.squareup.okhttp3:okhttp:4.9.1'
 	implementation 'com.github.ben-manes.caffeine:caffeine'
 	implementation 'org.apache.commons:commons-collections4:4.4'
+	runtimeOnly 'com.mysql:mysql-connector-j'
 }
