+++ src/main/java/com/zestbear/bitcoin/mybitcoin/service/UpbitAPI/Candle/CurrentValueAPI.java
@@ -1,7 +1,7 @@
 package com.zestbear.bitcoin.mybitcoin.service.UpbitAPI.Candle;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.zestbear.bitcoin.mybitcoin.domain.CurrentDataDto;
+import com.zestbear.bitcoin.mybitcoin.dto.CurrentDataDto;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
